Review: added AbstractTaskManagerTest
